+++ model/src/main/scala/org/broadinstitute/dsde/workbench/model/models.scala
@@ -1,0 +1,5 @@
+package org.broadinstitute.dsde.workbench.model
+
+import java.util.UUID
+
+final case class TraceId(uuid: UUID) extends AnyVal