+++ util/src/test/scala/org/broadinstitute/dsde/workbench/util/TestUtil.scala
@@ -2,6 +2,7 @@ package org.broadinstitute.dsde.workbench.util
 
 import cats.effect.{ContextShift, IO, Timer}
 import org.scalatest.Assertion
+import org.scalatest.prop.{Configuration, PropertyChecks}
 
 import scala.concurrent.Future
 import scala.concurrent.ExecutionContext.global
@@ -11,4 +12,8 @@ trait WorkbenchTest {
   implicit val cs: ContextShift[IO] = IO.contextShift(global)
 
   def ioAssertion(test: => IO[Assertion]): Future[Assertion] = test.unsafeToFuture()
+}
+
+trait PropertyBasedTesting extends PropertyChecks with Configuration {
+  implicit override val generatorDrivenConfig: PropertyCheckConfiguration = PropertyCheckConfiguration(minSuccessful = 3)
 }