+++ util/CHANGELOG.md
@@ -4,8 +4,10 @@ This file documents changes to the `workbench-util` library, including notes on 
 
 ## 0.5
 - add retry for IO
+- upgrade cats-effect
+- add `org.broadinstitute.dsde.workbench.util.PropertyBasedTesting`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.5-6942040"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.5-TRAVIS-REPLACE-ME"`
 
 ### Changed
 - Moved `SamModel`